+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTableCellRendererFactory.java
@@ -44,7 +44,7 @@ public class EntityTableCellRendererFactory implements FilterTableCellRendererFa
 	}
 
 	@Override
-	public TableCellRenderer tableCellRenderer(FilterTableColumn<Attribute<?>> column) {
+	public final TableCellRenderer tableCellRenderer(FilterTableColumn<Attribute<?>> column) {
 		return builder(column).build();
 	}